requirements.txt
yt-dlp
discord.py
nekos.py
websockets
beautifulsoup4